hw 6.4 done int

+++ src/Employee.java
@@ -1,4 +1,4 @@
 public interface Employee  {
 
-    double getMonthSalary();
+    int getMonthSalary();
 }
